requirements.txt
torch
torch-audiomentations
torch-pitch-shift
torchaudio
torchmetrics
torchvision
openai-whisper
colorama
ffutils
ffmpeg